Move inner classes of ExternalTextureFlutterActivity into their own pacakge.
