+++ testing/scenario_app/android/app/src/main/java/dev/flutter/scenarios/externaltextures/SurfaceRenderer.java
@@ -1,0 +1,13 @@
+package dev.flutter.scenarios.externaltextures;
+
+import android.view.Surface;
+
+import java.util.concurrent.CountDownLatch;
+
+public interface SurfaceRenderer {
+    void attach(Surface surface, CountDownLatch onFirstFrame);
+
+    void repaint();
+
+    void destroy();
+}
